volunteer-coordination-system/backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
PyMySQL==1.1.0
cryptography==42.0.2
python-dotenv==1.0.0
Werkzeug==3.0.1
